Add minimal integration test infrastructure to diesel_cli
This adds two small integration tests to CLI to demonstrate the style
that I'd like these tests to be written in.

As part of this, I went ahead and added a hidden `diesel database drop`
command so we could be sure that the database created for a given test
is removed when the test finishes running. To ensure that tests don't
have to worry about cleaning things like databases up, `Project` drops
the database automatically, and it is assumed that no databases would
have been created unless they are going through that struct.