Refactor tests added in #206
We don't care about the STDOUT output in this test. Additionally, rather
than coupling the test to how the database URL is being passed (which is
brittle and would lead to a false positive if that ever changed), we can
be more explicit about what we want when the command is created.

/cc @mcasper