Reverse the lifetime dependency of `#[insertable_into]`
I'm fairly certain this is a bug in Rust, not Diesel. The resulting
lifetime seems like it should be nonsense. We're now saying that the
reference to the struct must outlive one of the elements of the struct.
The fact that this only occurs when the lifetime `'a` appears more than
once supports that this is a Rust bug.

Regardless, this works around it, and everything seems to still be
passing.

Fixes #245