Add support for the `DISTINCT` keyword
A few things to note about the implementation. I've opted for 2 unit
structs instead of something like an enum or boolean so that the code
gets inlined via monomorphisation instead of containing a branch (e.g.
if you don't use distinct, it's free). As always, it's enforced that our
query builder doesn't do "real work" by
https://github.com/sgrif/diesel/blob/e7155c98d99996abb795c0bb736341ccf629930b/diesel_tests/tests/perf_details.rs#L8-L13

This ended up churning a lot of `select_dsl`. I've reformatted the code
surrounding what was touched to reduce churn in the future.

This doesn't yet add support for `DISTINCT ON(expr)`, as that is a PG
specific feature and can be added as a separate PR.

An alternative API for this would be something like
`users.select(distinct(name))`. I actually think I might prefer that
API, but actually enforcing correctness with it (making sure `distinct`
can only be passed to `select` and not something like `filter` --
disallowing things like `max(distinct(id))`, etc) would be pretty much
impossible.

Fixes #253.