Allow users to opt into usage of specialization
This is a change that I want to make eventually just to clean up our
code. However, there is a concrete case that can't wait for
specialization to land. There is not way for a user to implement
`FromSqlRow` for a nullable field of a custom type (e.g. a PG enum).

Using diesel with the `unstable` feature now uses specialization, which
we can start using elsewhere. We unfortunately still won't be able to
clean up the code base until specialization is stable, but we can at
least start to look at what the code *might* look like.