Release v0.6.0
In addition to a myriad of bug fixes, there are two headline features
for this release. We now provide an API for dealing with boxed queries.
Through our usage of Diesel for web apps, we found that this was a
feature we needed for things like complex search forms where you may
want to conditionally modify a query. For details, check out
http://docs.diesel.rs/diesel/prelude/trait.BoxedDsl.html

Additionally, this release fleshes out our support for SQLite by
allowing the `infer_schema!` macro to be used with it. Thanks to
@weiznich for working on this.

There were many other changes that weren't mentioned here. For the full
changelog, please see
https://github.com/sgrif/diesel/blob/master/CHANGELOG.md#050-2016-04-12