Execute Benchmarks for 0, 1, 10, 100, 1k, 10k Rows
This moves benchmark code into macros to easily abstract over the number
of rows to benchmark. Because you can't use macros in function name
position, we have to write each benchmark name ourselves. (I tried
moving the fns into a named mod, but there were some visibility issues
and I gave up on that.)

For the numbers in the benchmark names to align, they are padded with
several underscores. For this to not print warning,
`#![allow(non_snake_case)]` was added.