Add support for comparison operators when used with PG arrays
These were all fairly straightforward. We used methods for all three,
even though one maps to the `&&` operator in PG. This is because we
can't overload the return type of `&&` in Rust, and we probably wouldn't
do that even if we could because I think `&&` is an abysmally unclear
operator for "overlap"

We had to add a dummy `users` table because of the call to `Insertable!`
in the doctest setup file. We can't just do it there, because a few
examples need a different schema. I'm not sure how I want to restructure
it yet, but those will be removed in a future commit.