Give an error message when an invalid PG url is given to CLI
Right now we assume URLs that aren't for PG are for SQLite instead.
However, the CLI crate can be compiled without SQLite support. In this
case, we end up hitting the `unreachable` branch in our match statement.
I'd like to eventually change this code to use an enum, similar to
codegen. In the short term though, we just need to improve the error
message.

Fixes #310