Don't strip `Option` when parsing struct bodies
While some macros like `Insertable!` want to treat option separately,
not all macros do. Since macros can't appear in a type position, it is
painful for those macros to add the `Option` back later. It seems that
all cases where we could just omit the `Option` entirely in whatever
associated type we're building are able to handle it fine, as they're
most often dealing with an `AsExpression` constraint