Allow trailing commas to be omitted from struct definitions
Particularly with tuple structs, it's fairly common to not have a
trailing comma. As our parsing macro is written today, we are enforcing
that all struct definitions have a trailing comma.

This is actually somewhat tricky to handle in a way that doesn't
duplicate half of our parsing code. This is the most factored way to
handle it that I could find. We blindly add an additional comma to the
end of the body of all struct definitions during parsing. This means
that if there *was* a trailing comma, our parser will eventually consume
everything except a single token. Therefore, we can just match exactly
that case and throw the final comma away.