Implement `AsExpression` for `&&[T]`
Typically we implement this trait for `T -> ST`, `T -> Nullable<ST>`,
`&T -> ST`, and `&T -> Nullable<ST>` using the `expression_impls!`
macro. When the type itself is a reference, we need to do this for a
double reference as well (for example, `&&str` implements
`AsExpression<Text>`). The `expression_impls!` macro isn't used for
array types, as it has an additional type parameter, and as such I
forgot the `&&[T]` case.

Fixes #308