Move the trait bounds for `FindDsl` onto the impl (#352)
Since Rust does not infer the where clasue of traits when using that
trait as a bound, you effectively have to restate the entire where
clause every time you want to reference that trait. This makes `FindDsl`
basically impossible to actually abstract over in practice.

This promotes it to the same place that the rest of the query DSL is,
where it has an abstract output and a helper type to follow the return
type chain through multiple levels.

This is part of a larger audit of this module to ensure consistency and
ease of abstraction throughout.