Provide the ability to handle unique constraint violations
This lays the groundwork for a more robust error handling story, but to
start we are only handling `UniqueViolation` specifically. SQLite
separates primary key violations from unique constraints, but I've opted
to group them together.

Due to how PG handles errors, we will only know one of the constraints
that was violated (if more than one was violated), and you can only get
the constraint name, not the column name (in the single column case).
SQLite provides none of this information, and the error message would
have to be parsed to determine which constraint failed.

Ultimately I think all of this is fine, as you rarely are expecting more
than exactly one constraint to fail in a given query.

Fixes #360