Handle optional foreign keys in `belongs_to`
This fixes the test failures caused by the previous commit. The
procedural macro had some behavior that the stable macro does not, where
it skips the implementation of `BelongsTo` if the foreign key field is
an option. I think what I'd like to eventually do is change this so
that if the foreign key is optional, you write
`#[belongs_to(Option<User>)]` instead. I cannot mirror the procedural
macro behavior in the stable macro, as I there is no way to match "two
patterns which are the same".

I've also brought back the error behavior of the procedural version of
`Identifiable`, as the stable macro gives an unhelpful error message if
you place it on a struct with no id field.