Allow tuples to be used with `grouped_by`
The current implementation of associations basically ignores multiple
level nesting. It lets us get `Vec<(User, Vec<Post>)>`, but not
`Vec<(User, Vec<(Post, Vec<Comment>)>)>`. This allows a tuple of
arbitrary size to be used with `grouped_by`, as long as the first
element in the tuple is the actual child used in the association. This
means that `Vec<(Post, Vec<Comment>)>` implements `GroupedBy<User>`.
I've also moved the impl off of `Vec` and onto `T: IntoIterator`, to get
rid of some unneccesary `collect` calls in tests.