Support nullable foreign keys in associations
The `BelongsTo!` macro now does crazy hacks to pull out the foreign key field
from the list given, checks if it's an option type, and will generate different
code as a result. The types that `BelongingToDsl` and `GroupedBy` get
implemented for are the same as if the foreign key was not nullable.
`GroupedBy` will simply discard any records with `None` for the foreign key
(which it is unlikely to receive in the first place since the result of
`BelongingToDsl` implicitly never returns records with a null value for the
foreign key.)

This differs from the direction I had previously discussed, where we would
implement the various traits for `Option<Parent>`. There are two main reasons
for this:

- The impl of `BelongingToDsl` would be gross, as we now have to care about
  `NULL` appearing anywhere and give it special handling.
- In the real world, you're never dealing with "records belonging to these 3
  parents, and records which have no parent".

For simplication in the code, I've set up `BelongsTo` to always return an
option for the foreign key. This causes an unneccesary conditional to appear in
the body of `GroupedBy`. However, I'm fairly certain that conditional will get
optimized away after inlining. I have not checked the generated ASM or LLVM IR
to confirm this.