Remove the `Column` constraint from `PrimaryKey`
We cannot have that constraint in order to have composite primary keys.
It might make sense to introduce some sort of trait that allows us to
guarantee that this type is always either a column or a tuple that only
contains columns, but at the end of the day we're only using this to
generate things like `FindDsl` so any selectable expression should do
here.